# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Banco Demo API",
    "version": "1.0.0",
    "description": "Backend bancário demo (contrato-compartilhado v1.0). Sessão por cookie bank_session. Transferências assíncronas via Saga: POST → PENDING/PROCESSING → GET → COMPLETED/FAILED."
  },
  "servers": [
    {
      "url": "http://127.0.0.1:3001",
      "description": "Backend local"
    }
  ],
  "paths": {
...
    "/v1/auth/signup": {
      "post": {
        "summary": "Cadastrar usuário + conta + sessão",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SignupBody"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Criado + cookie bank_session (Set-Cookie)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResult"
                }
              }
            },
            "headers": {
              "Set-Cookie": {
                "description": "bank_session HttpOnly;SameSite=Lax;Path=/;Max-Age=86400",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "VALIDATION_ERROR",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "403": {
            "description": "ORIGIN_NOT_ALLOWED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "409": {
            "description": "EMAIL_ALREADY_EXISTS",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "INTERNAL_ERROR",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/auth/signin": {
      "post": {
        "summary": "Entrar (cria sessão)",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SigninBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK + cookie bank_session",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResult"
                }
              }
            },
            "headers": {
              "Set-Cookie": {
                "description": "bank_session",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "VALIDATION_ERROR",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "401": {
            "description": "INVALID_CREDENTIALS",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "403": {
            "description": "ORIGIN_NOT_ALLOWED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "INTERNAL_ERROR",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/auth/signout": {
      "post": {
        "summary": "Sair (idempotente, sempre 204)",
        "requestBody": {
          "required": false,
          "description": "Sem corpo; Content-Type json com corpo vazio é aceito."
        },
        "responses": {
          "204": {
            "description": "Saiu + cookie expirado"
          },
          "403": {
            "description": "ORIGIN_NOT_ALLOWED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "INTERNAL_ERROR",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/recipients/{accountId}": {
      "get": {
        "summary": "Consultar destinatário (só nome + ID)",
        "parameters": [
          {
            "name": "accountId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Recipient"
                }
              }
            }
          },
          "401": {
            "description": "UNAUTHENTICATED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "RECIPIENT_NOT_FOUND",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "422": {
            "description": "SELF_RECIPIENT",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "INTERNAL_ERROR",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/contacts": {
      "get": {
        "summary": "Listar contatos (nickname, id)",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContactsList"
                }
              }
            }
          },
          "401": {
            "description": "UNAUTHENTICATED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "INTERNAL_ERROR",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Salvar contato",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddContactBody"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Criado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Contact"
                }
              }
            }
          },
          "400": {
            "description": "VALIDATION_ERROR",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "401": {
            "description": "UNAUTHENTICATED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "403": {
            "description": "ORIGIN_NOT_ALLOWED",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "RECIPIENT_NOT_FOUND",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "409": {
            "description": "CONTACT_ALREADY_EXISTS",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "422": {
            "description": "SELF_RECIPIENT",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "INTERNAL_ERROR",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bankSession": {
        "type": "apiKey",
        "in": "cookie",
        "name": "bank_session",
        "description": "Sessão opaca HttpOnly, 24h. Signup/signin emitem Set-Cookie; signout expira."
      }
    },
    "schemas": {
...
      "AuthResult": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "user",
          "account"
        ],
        "properties": {
          "user": {
            "type": "object",
            "additionalProperties": false,
            "required": [
              "id",
              "name",
              "email",
              "createdAt"
            ],
            "properties": {
              "id": {
                "type": "string"
              },
              "name": {
                "type": "string"
              },
              "email": {
                "type": "string"
              },
              "createdAt": {
                "type": "string"
              }
            }
          },
          "account": {
            "type": "object",
            "additionalProperties": false,
            "required": [
              "id",
              "currency",
              "balanceCents"
            ],
            "properties": {
              "id": {
                "type": "string"
              },
              "currency": {
                "type": "string",
                "const": "BRL"
              },
              "balanceCents": {
                "type": "integer"
              }
            }
          }
        }
      },
...
      "Recipient": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "accountId",
          "name"
        ],
        "properties": {
          "accountId": {
            "type": "string"
          },
          "name": {
            "type": "string"
          }
        }
      },
      "Contact": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "id",
          "nickname",
          "recipientAccountId",
          "recipientName",
          "createdAt"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "nickname": {
            "type": "string"
          },
          "recipientAccountId": {
            "type": "string"
          },
          "recipientName": {
            "type": "string"
          },
          "createdAt": {
            "type": "string"
          }
        }
      },
      "ContactsList": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "items"
        ],
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "type": "object",
              "additionalProperties": false,
              "required": [
                "id",
                "nickname",
                "recipientAccountId",
                "recipientName",
                "createdAt"
              ],
              "properties": {
                "id": {
                  "type": "string"
                },
                "nickname": {
                  "type": "string"
                },
                "recipientAccountId": {
                  "type": "string"
                },
                "recipientName": {
                  "type": "string"
                },
                "createdAt": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
...
      "ApiError": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "error",
          "requestId"
        ],
        "properties": {
          "error": {
            "type": "object",
            "additionalProperties": false,
            "required": [
              "code",
              "message"
            ],
            "properties": {
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string"
              },
              "details": {
                "type": "array",
                "items": {
                  "type": "object",
                  "additionalProperties": false,
                  "required": [
                    "field",
                    "message"
                  ],
                  "properties": {
                    "field": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "requestId": {
            "type": "string"
          }
        }
      },
      "SignupBody": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "name",
          "email",
          "password"
        ],
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 80
          },
          "email": {
            "type": "string",
            "minLength": 1,
            "maxLength": 254
          },
          "password": {
            "type": "string",
            "minLength": 8,
            "maxLength": 72
          }
        }
      },
      "SigninBody": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "email",
          "password"
        ],
        "properties": {
          "email": {
            "type": "string",
            "minLength": 1,
            "maxLength": 254
          },
          "password": {
            "type": "string",
            "minLength": 8,
            "maxLength": 72
          }
        }
      },
      "AddContactBody": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "nickname",
          "recipientAccountId"
        ],
        "properties": {
          "nickname": {
            "type": "string",
            "minLength": 1,
            "maxLength": 60
          },
          "recipientAccountId": {
            "type": "string",
            "minLength": 1
          }
        }
      },
...
    }
  },
  "security": [
    {
      "bankSession": []
    }
  ]
}